refactor(engine): split walk_mut, drop redundant Arc wraps, per-backend modules
Three structural cleanups that tighten the engine surface without
changing observable behavior:

1. walk_mut → root_mut() + embeds_mut()
   - DocumentTree::walk_mut (HRTB-callback shape) replaced by a sync
     pair: root_mut() returns the root NodeMut, embeds_mut() yields
     nested ones. Phases write: dispatch(root_mut()).await?; for node
     in embeds_mut() { dispatch(node).await? }.
   - Drops per-node Box::pin allocations, per-node Arc clones, and the
     for<'a> FnMut HRTB workaround. Phases borrow stack-local config
     directly.
   - The for<'a> FnMut + Send + 'a workaround was forced by a known
     compiler limitation (rust-lang/rust#70263, #109924) and tracked
     RTN/AsyncFnMut work that's stalled on stable. The new shape
     doesn't fight the language.

2. Drop redundant Arc<ExtractionEngine|DetectionEngine|RedactionConfig>
   wraps below EngineInner
   - EngineInner stays Arc-wrapped (genuine singleton, shared across
     Engine clones). Below that there's no sharing: Pipeline,
     RunContext, and each phase struct hold engines by value.
   - DetectionEngine gains #[derive(Clone)]; ExtractionEngine and
     RedactionConfig already derive Clone. Pipeline::execute clones
     once per run into RunContext; from_context clones once per phase.
     Each clone is a few atomic increments — the engines hold Arc'd
     services internally.
   - Engine accessors drop one layer of indirection: &Arc<E> → &E.

3. Extraction restructure: per-backend modules instead of per-modality
   - extraction/{audio,image,tabular,text}/ → folded into engine.rs;
     the per-modality populators (populate_text_blocks,
     populate_image_embeds, populate_tabular_blocks,
     populate_image_doc, populate_audio_doc) and the dispatch
     switchboard are now &self methods on ExtractionEngine.
   - extraction/image/ocr.rs → ocr/mod.rs (folder module).
   - extraction/audio/stt.rs → stt/mod.rs (folder module).
   - Mirrors detection/'s by-recognizer-kind layout.

Detection got the matching shape: detect_image and dispatch are now
&self methods on DetectionEngine. The free functions in detection/mod.rs
are gone; DetectionPhase::apply calls self.engine.dispatch(...).

Also drops a few cosmetic bindings (`let engine = &self.engine;`)
that just shadowed self.field, and the dead `let _ = &plan.X;` lines
that suppressed unused-field warnings — the populators now take only
the plan slices they use.

Workspace builds clean; nvisy-engine lib tests 68/68 green; fmt
clean; clippy --workspace --all-targets clean.
